# Conflicts:
#	tpf-service-autoallocation/src/main/java/vn/com/tpf/microservices/dao/UserDetailsDAO.java
#	tpf-service-autoallocation/src/main/java/vn/com/tpf/microservices/services/AutoAllocationService.java

+++ tpf-service-autorouting/src/main/java/vn/com/tpf/microservices/dao/LogChoiceRoutingDAO.java
@@ -5,5 +5,5 @@ import vn.com.tpf.microservices.models.LogChoiceRouting;
 
 public interface LogChoiceRoutingDAO extends JpaRepository<LogChoiceRouting, Long> {
 
-    LogChoiceRouting findByIdLog(int idLog);
+    LogChoiceRouting findByIdLog(Long idLog);
 }
